refactor: extract sendProgress helper into mcp/progress.ts
Break the circular dependency where tools imported from their
own registration module (mcp/index.ts -> tools -> mcp/index.ts).
