+++ src/mcp/index.ts
@@ -1,35 +1,13 @@
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { RequestHandlerExtra } from "@modelcontextprotocol/sdk/shared/protocol.js";
-import type { ServerNotification, ServerRequest } from "@modelcontextprotocol/sdk/types.js";
 import { FigmaService, type FigmaAuthOptions } from "../services/figma.js";
 import { Logger } from "../utils/logger.js";
+import type { ToolExtra } from "./progress.js";
 import {
   downloadFigmaImagesTool,
   getFigmaDataTool,
   type DownloadImagesParams,
   type GetFigmaDataParams,
 } from "./tools/index.js";
-
-export type ToolExtra = RequestHandlerExtra<ServerRequest, ServerNotification>;
-
-/**
- * Send a progress notification if the client requested one via progressToken.
- * No-ops silently when the client didn't ask for progress.
- */
-export async function sendProgress(
-  extra: ToolExtra,
-  progress: number,
-  total?: number,
-  message?: string,
-): Promise<void> {
-  const progressToken = extra._meta?.progressToken;
-  if (progressToken === undefined) return;
-
-  await extra.sendNotification({
-    method: "notifications/progress",
-    params: { progressToken, progress, total, message },
-  });
-}
 
 const serverInfo = {
   name: "Figma MCP Server",
