+++ src/mcp/progress.ts
@@ -1,0 +1,20 @@
+import type { RequestHandlerExtra } from "@modelcontextprotocol/sdk/shared/protocol.js";
+import type { ServerNotification, ServerRequest } from "@modelcontextprotocol/sdk/types.js";
+
+export type ToolExtra = RequestHandlerExtra<ServerRequest, ServerNotification>;
+
+/** No-ops silently when the client didn't ask for progress (no progressToken). */
+export async function sendProgress(
+  extra: ToolExtra,
+  progress: number,
+  total?: number,
+  message?: string,
+): Promise<void> {
+  const progressToken = extra._meta?.progressToken;
+  if (progressToken === undefined) return;
+
+  await extra.sendNotification({
+    method: "notifications/progress",
+    params: { progressToken, progress, total, message },
+  });
+}
